+++ docs/files/ros2-action-client.py
@@ -5,25 +5,25 @@ import time
 
 global result
 
-client = roslibpy.Ros(host='localhost', port=9090)
+client = roslibpy.Ros(host="localhost", port=9090)
 client.run()
 
 action_client = roslibpy.ActionClient(client,
-                                      '/fibonacci',
-                                      'custom_action_interfaces/action/Fibonacci')
+                                      "/fibonacci",
+                                      "custom_action_interfaces/action/Fibonacci")
 result = None
 
 def result_callback(msg):
     global result
-    result = msg['result']
+    result = msg["result"]
 
 def feedback_callback(msg):
-    print('Action feedback:',msg['partial_sequence'])
+    print(f"Action feedback: {msg['partial_sequence']}")
 
 def fail_callback(msg):
-    print('Action failed:',msg)
+    print(f"Action failed: {msg}")
 
-goal_id = action_client.send_goal(roslibpy.ActionGoal({'order': 8}),
+goal_id = action_client.send_goal(roslibpy.ActionGoal({"order": 8}),
                                   result_callback,
                                   feedback_callback,
                                   fail_callback)
@@ -31,4 +31,4 @@ goal_id = action_client.send_goal(roslibpy.ActionGoal({'order': 8}),
 while result == None:
     time.sleep(1)
     
-print('Action result: {}'.format(result['sequence']))
+print("Action result: {}".format(result["sequence"]))