+++ src/roslibpy/__init__.py
@@ -41,13 +41,13 @@ as an argument to their constructors.
 Main ROS concepts
 =================
 
-ROS1 vs ROS2
+ROS 1 vs ROS 2
 ------------
 
-This library has been tested to work with ROS1. ROS2 should work, but it is still
+This library has been tested to work with ROS 1. ROS 2 should work, but it is still
 in the works.
 
-One area in which ROS1 and ROS2 differ is in the header interface. To use ROS2, use
+One area in which ROS 1 and ROS 2 differ is in the header interface. To use ROS 2, use
 the header defined in the `roslibpy.ros2` module.
 
 .. autoclass:: roslibpy.ros2.Header
@@ -85,7 +85,7 @@ model via :class:`Services <Service>`.
 Actions
 --------
 
-An Action client for ROS2 Actions can be used by managing goal/feedback/result
+An Action client for ROS 2 Actions can be used by managing goal/feedback/result
 messages via :class:`ActionClient <ActionClient>`.
 
 .. autoclass:: ActionClient