+++ CHANGELOG.rst
@@ -12,11 +12,11 @@ Unreleased
 
 **Added**
 
+* Added a ROS2-compatible header class in ``roslibpy.ros2.Header``.
+
 **Changed**
 
 **Fixed**
-
-* Removed seq field from the header message type, as this is no longer supported in ROS2
 
 **Deprecated**