+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Unreleased
 ----------
 
 **Added**
+* Added ROS 2 action client object with limited capabilities ``roslibpy.ActionClient``.
 
 * Added websocket header support to the ROS-client.
 
@@ -27,7 +28,7 @@ Unreleased
 
 **Added**
 
-* Added a ROS2-compatible header class in ``roslibpy.ros2.Header``.
+* Added a ROS 2 compatible header class in ``roslibpy.ros2.Header``.
 
 **Changed**